+++ example/index.js
@@ -1,4 +1,8 @@
-import { PrivateKey, decrypt, encrypt } from "eciesjs";
+import { ECIES_CONFIG, PrivateKey, decrypt, encrypt } from "eciesjs";
+import { Buffer } from "node:buffer";
+
+globalThis.Buffer = Buffer;
+ECIES_CONFIG.symmetricNonceLength = 12
 
 const sk = new PrivateKey()
 const data = Buffer.from("hello world🌍")