+++ example/vite.config.ts
@@ -1,5 +1,17 @@
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  plugins: [],
+  build: {
+    rollupOptions: {
+      output: {
+        manualChunks(id) {
+          if (id.includes("@noble")) {
+            return "noble";
+          } else if (id.includes("buffer")) {
+            return "buffer";
+          }
+        },
+      },
+    },
+  },
 });