+++ src/utils/symmetric.ts
@@ -1,8 +1,8 @@
-import { xchacha20poly1305 as xchacha20 } from "@noble/ciphers/chacha";
 import { Cipher, concatBytes } from "@noble/ciphers/utils";
 import { randomBytes } from "@noble/ciphers/webcrypto";
 
-import { aes256cbc, aes256gcm } from "@ecies/ciphers";
+import { aes256cbc, aes256gcm } from "@ecies/ciphers/aes";
+import { xchacha20 } from "@ecies/ciphers/chacha";
 import { symmetricAlgorithm, symmetricNonceLength } from "../config";
 import { AEAD_TAG_LENGTH, XCHACHA20_NONCE_LENGTH } from "../consts";